+++ src/app/app.module.ts
@@ -8,6 +8,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SharedModule } from './shared/shared.module';
 import { ProductsModule } from './products/products.module';
 import { CartModule } from './cart/cart.module';
+import { OrdersModule } from './orders/orders.module';
 
 @NgModule({
   declarations: [AppComponent, FirstComponent],
@@ -17,6 +18,7 @@ import { CartModule } from './cart/cart.module';
     BrowserAnimationsModule,
     ProductsModule,
     CartModule,
+    OrdersModule,
     SharedModule,
   ],
   providers: [],