+++ src/app/cart/components/cart/cart.component.ts
@@ -1,26 +1,33 @@
-import { Component, Input, EventEmitter, Output } from '@angular/core';
+import {
+  Component,
+  Input,
+  EventEmitter,
+  Output,
+  ChangeDetectionStrategy,
+} from '@angular/core';
 import { CartModel } from '../../models/cart.model';
 
 @Component({
   selector: 'swm-cart',
   templateUrl: './cart.component.html',
   styleUrls: ['./cart.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CartComponent {
   @Input() product!: CartModel;
   @Output() delete = new EventEmitter<number>();
-  @Output() plus = new EventEmitter<number>();
-  @Output() minus = new EventEmitter<number>();
+  @Output() quantityIncrease = new EventEmitter<number>();
+  @Output() quantityDecrease = new EventEmitter<number>();
 
   onDelete(id: number) {
     this.delete.emit(id);
   }
 
-  onPlus(id: number) {
-    this.plus.emit(id);
+  onQuantityIncrease(id: number) {
+    this.quantityIncrease.emit(id);
   }
 
-  onMinus(id: number) {
-    this.minus.emit(id);
+  onQuantityDecrease(id: number) {
+    this.quantityDecrease.emit(id);
   }
 }