+++ src/app/cart/models/cart.model.ts
@@ -1,5 +1,6 @@
 import { ProductModel } from 'src/app/products/models/product.model';
 
 export interface CartModel extends ProductModel {
-  productCount: number;
+  totalQuantity: number;
+  totalCost: number;
 }