+++ src/app/first/first.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { ProductModel } from '../products/models/product.model';
 import { ProductsMock } from '../products/mocks/product.mock';
 
@@ -6,6 +6,7 @@ import { ProductsMock } from '../products/mocks/product.mock';
   selector: 'swm-first',
   templateUrl: './first.component.html',
   styleUrls: ['./first.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class FirstComponent {
   products: Array<ProductModel> = ProductsMock;