+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ProductsService } from '../../services/products.service';
 import { ProductModel } from '../../models/product.model';
 import { CartService } from 'src/app/cart/services/cart.service';
@@ -7,6 +7,7 @@ import { CartService } from 'src/app/cart/services/cart.service';
   selector: 'swm-product-list',
   templateUrl: './product-list.component.html',
   styleUrls: ['./product-list.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductListComponent implements OnInit {
   products!: Array<ProductModel>;
@@ -20,7 +21,7 @@ export class ProductListComponent implements OnInit {
     this.products = this.productsService.getProducts();
   }
 
-  onBuy(product: ProductModel) {
+  buy(product: ProductModel) {
     this.cartService.addNewProduct(product);
   }
 }