+++ src/app/products/components/product/product.component.ts
@@ -1,10 +1,17 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  EventEmitter,
+  Input,
+  Output,
+} from '@angular/core';
 import { ProductModel } from '../../models/product.model';
 
 @Component({
   selector: 'swm-product',
   templateUrl: './product.component.html',
   styleUrls: ['./product.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductComponent {
   @Input() product!: ProductModel;