+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,17 @@
+import { Directive, HostBinding, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[swmHighlight]',
+})
+export class HighlightDirective {
+  @HostBinding('style.background-color') backgroundColor = 'white';
+  @HostListener('mouseenter')
+  mouseover() {
+    this.backgroundColor = '#99EDC3';
+  }
+
+  @HostListener('mouseleave')
+  mouseleave() {
+    this.backgroundColor = 'white';
+  }
+}